make source hover rather than click
I love the way you added the source functionality here!! This is just a suggestion – but I think it might work better if source appears when someone hovers their cursor over the "source" label ... rather than clicks it. (I think on a touch screen it will work that if you touch "source" it will trigger, and when you touch anywhere else it will revert... but we'd need to test to make sure.) Also from a design standpoint, I think lowercase "source" looks better.